+++ sdk/education/azure-resourcemanager-education/src/main/java/com/azure/resourcemanager/education/EducationManager.java
@@ -226,7 +226,7 @@ public final class EducationManager {
                 .append("-")
                 .append("com.azure.resourcemanager.education")
                 .append("/")
-                .append("1.0.0-beta.1");
+                .append("1.0.0-beta.2");
             if (!Configuration.getGlobalConfiguration().get("AZURE_TELEMETRY_DISABLED", false)) {
                 userAgentBuilder
                     .append(" (")